GNUmakefile: support alternate library installation destinations
Prior to this commit, the libs3 shared object and static archive would
always be installed into ${DESTDIR}/lib/.

This works fine for some distributions, but on Red Hat x86_64, we need
to be able to install into /usr/lib64 instead of /usr/lib.

Allow the configuration of the library installation via the LIBDIR
variable.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>