Fix corrupted request time headers due to thread-unsafe gmtime
gmtime() is used to convert the system time to GMT so that the S3
request time can be appended to the headers.  It returns a pointer
to a library static struct tm.  This works fine in the single-
threaded case, but can cause request times to be corrupted should
libs3 be called in a multithreaded use case.

Replaces the gmtime() with a gmtime_r() call which takes a pointer
to a thread-local struct tm, fixing the issue.