Fix unitialized hostHeader element
hostHeader, when not used with virtual host style URIs, would never
be set to any value.  This would result in occasional, random headers
being sent to S3 depending on what was left over in memory.

Just set to the \0 string in normal case, this way we don't add it
or garbage ourselves.