Issue #36 - Expires header is set incorrectly
The "Expires" field is not needed for a create bucket request and
in fact, setting it to zero can cause some S3 implementations to
fail the request because well, it has expired.

Therefore, we set the "expires" S3PutParameter of the create bucket
call to -1 in order to bypass generating this HTTP header entirely.