Use git tags for dist tar.gz versioning
In order to provide a much cleaner release build (RPM, tar.gz, etc)
mechanism, it is highly desirable to use git tags to drive the
versioning. This allows for much cleaner automation (git tag && git push
--tags) along with providing git tracking of these release artifact
versions.

We are taking the idea of GIT-VERSION-GEN from the upstream git
repository itself and using the dynamic GNUMakefile include of the
generated data. This populates VERSION, RELEASE and FULL_VERSION from
the latest git tag. This requires the initial tags be populated into the
repository before trying to use this change.

The latest EPEL RPM set could be tagged like this:
$ git tag -a v2.0-0.2.20150902 247ba1b

We are also adding a 'make dist' target to dump a .tar.gz that could be
used for RPM or deb builds in the future. This makes full use of the
versioning information in the GIT-VERSION-FILE and ensures that the
library is built with the same versions as are constructed in release
artifacts.