VSM-1178 update spec to be RHEL agnostic
On RHEL7, we need to move install to lib64 as that is the correct spot.
We can use the _libdir macro to get RPM to figure this out for us across
the different OS versions.

We also need to chmod the .so, as those permissions is how the automatic
find-provides process picks up the .so and adds our library to the
dependencies we provide.