Incomplete listing on Backblaze S3
Contrary to AWS, Backblaze adds `IsTruncated` (and others) only after the `CommonPrefixes` and `Contents`. It's pagination is 1000, what is a multiple of libs3's `MAX_COMMON_PREFIXES`. So if you have prefixes-only listing, you get 125 callbacks with 32 `commonPrefixes` in each. None of them will have valid `isTruncated` yet. And in `listBucketCompleteCallback`, where `isTruncated` is finally known, both `contentsCount` and `commonPrefixesCount` will be zero, and the callback will not be triggered.